+++ scripts/test.js
@@ -3,7 +3,7 @@
 const jest = require('jest')
 const createJestConfig = require('../config/createJestConfig')
 
-function test(args) {
+async function test(args) {
   // Watch unless on CI, in coverage mode, or explicitly running all tests
   if (!process.env.CI && args.indexOf('--coverage') === -1) {
     args.push('--watch')